// backend/src/modules/production-orders/dto/update-production-order-volume-check.dto.ts
export class UpdateProductionOrderVolumeCheckDto {
  package_type?: string | null;
  requirement?: string | null;
  dosage_form_stage?: string | null;
  unit_1_volume?: number | string | null;
  unit_2_volume?: number | string | null;
  unit_3_volume?: number | string | null;
  unit_4_volume?: number | string | null;
  unit_5_volume?: number | string | null;
  unit_6_volume?: number | string | null;
}